Fix complex sorting bug
Axis sorting order worked completely due to chance with the tilemap.

ISSUE: Foreground tilemap was rendered in one giant chunk. This means
that the GPU only has to use 1 instruction to draw every tile in the
foreground, but the drawing order for each tile was not preserved. The
sorting order for all tiles were rendered using the center of the entire
foreground map as its sorting point, but due to the structure of the
map, this did not cause issues.

SOLUTION: Render each tile in the foreground individually. This will
reduce performance if the foreground is too big, so object that are
consistently in the background / foreground should be put on a separate
layer with chunk rendering. This fixes the issue, but now there's a new
bug; When standing in front of a wall that normally would encase the
player, the player clips through the top tile if the wall is two tiles
high.